refactor(src/*): 修改数据类型并增加注释与测试
此次变更主要涉及多个文件中的数据类型的调整，将部分 `u64` 类型改为 `u32` 和 `u16` 类型，同时增加了相关函数的注释说明以及单元测试。这些改变旨在优化内存使用并提高代码的可读性及测试覆盖率。

+++ src/error.rs
@@ -22,7 +22,7 @@ pub enum PDFError {
     #[error("{0}")]
     PDFParseError0(String),
     #[error("Xref entry:({0},{1}) not found")]
-    XrefEntryNotFound(u64,u64),
+    XrefEntryNotFound(u32,u16),
     #[error("{0}")]
     ObjectAttrMiss(&'static str),
     #[error("End of file error")]
